feat target

+++ proj/src/model/crosshair.h
@@ -4,12 +4,12 @@
 #include <lcom/lcf.h>
 
 typedef struct {
-    int x, y;
-    int width, height;
+    int16_t x, y;
+    uint16_t width, height;
 } Crosshair;
 
-void crosshair_init(int x, int y);
-void crosshair_update_position(int delta_x, int delta_y);
+void crosshair_init(int16_t x, int16_t y);
+void crosshair_update_position(int16_t delta_x, int16_t delta_y);
 
 #endif
